refactor: linux mitm
